requirements.txt
langchain==0.1.13
langgraph==0.0.40
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
langchain-google-genai==0.0.7
google-generativeai>=0.3.0
